Remove compilation warnings.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@637 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ samples/c/facedetect.c
@@ -136,7 +136,9 @@ void detect_and_draw( IplImage* img )
 {
     double scale = 1.3;
     IplImage* gray = cvCreateImage( cvSize(img->width,img->height), 8, 1 );
-    IplImage* small_img = cvCreateImage( cvSize(img->width/scale,img->height/scale), 8, 1 );
+    IplImage* small_img = cvCreateImage( cvSize( (int)(img->width/scale),
+						 (int)(img->height/scale)),
+					 8, 1 );
     CvPoint pt1, pt2;
     int i;
 
@@ -156,10 +158,10 @@ void detect_and_draw( IplImage* img )
         for( i = 0; i < (faces ? faces->total : 0); i++ )
         {
             CvRect* r = (CvRect*)cvGetSeqElem( faces, i );
-            pt1.x = r->x*scale;
-            pt2.x = (r->x+r->width)*scale;
-            pt1.y = r->y*scale;
-            pt2.y = (r->y+r->height)*scale;
+            pt1.x = (int)(r->x*scale);
+            pt2.x = (int)((r->x+r->width)*scale);
+            pt1.y = (int)(r->y*scale);
+            pt2.y = (int)((r->y+r->height)*scale);
             cvRectangle( img, pt1, pt2, CV_RGB(255,0,0), 3, 8, 0 );
         }
     }
